+++ packages/blaxel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/blaxel
 
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies [e3ed89b]
+  - computesdk@2.0.2
+  - @computesdk/provider@1.0.22
+
 ## 1.5.0
 
 ### Minor Changes