+++ packages/daytona/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/daytona
 
+## 1.7.9
+
+### Patch Changes
+
+- Updated dependencies [e3ed89b]
+  - computesdk@2.0.2
+  - @computesdk/provider@1.0.22
+
 ## 1.7.8
 
 ### Patch Changes