+++ packages/render/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/render
 
+## 1.2.9
+
+### Patch Changes
+
+- Updated dependencies [e3ed89b]
+  - computesdk@2.0.2
+  - @computesdk/provider@1.0.22
+
 ## 1.2.8
 
 ### Patch Changes